requirements.txt
openai
resend
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
